Auto stash before checking out "origin/1-create-schema"
